divide render from element_commands
